requirements.txt
python-gitlab==3.1.0
